Add programs
Programs actually work, would you belive it? But easing was not implemented.

+++ CaptivePortalJas/VhPwm.h
@@ -13,7 +13,7 @@ class VhPwm{
     
   protected:
     int _channel;
-    int _duty;
+    float _duty;
